scripts/requirements.txt
scholarly>=1.7.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pyyaml>=6.0

